requirements-3.11.txt
# NOTE - this file is automatically generated during CICD builds, do not update manually
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    ./bin/build.sh gen_requirements 3.11
#
accessible-pygments==0.0.4
    # via pydata-sphinx-theme
alabaster==0.7.13
    # via sphinx
amalgam-lang==2.3.18
    # via howso-engine-api
babel==2.12.1
    # via
    #   pydata-sphinx-theme
    #   sphinx
beautifulsoup4==4.12.2
    # via pydata-sphinx-theme
certifi==2023.7.22
    # via
    #   howso-engine
    #   howso-engine-api
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
contourpy==1.1.0
    # via matplotlib
cryptography==41.0.3
    # via howso-engine-api
cycler==0.11.0
    # via matplotlib
deprecation==2.1.0
    # via howso-engine-api
docutils==0.19
    # via
    #   myst-parser
    #   pydata-sphinx-theme
    #   sphinx
faker==19.3.1
    # via howso-engine-api
fonttools==4.42.1
    # via matplotlib
howso-engine==1.1.46
    # via -r requirements.in
howso-engine-api==7.1.0
    # via
    #   -r requirements.in
    #   howso-engine
howso-openapi-client==24.0.15
    # via howso-engine-api
humanize==4.8.0
    # via howso-engine-api
idna==3.4
    # via requests
imagesize==1.4.1
    # via sphinx
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   myst-parser
    #   sphinx
joblib==1.3.2
    # via scikit-learn
kiwisolver==1.4.5
    # via matplotlib
markdown-it-py==2.2.0
    # via
    #   mdit-py-plugins
    #   myst-parser
    #   rich
markupsafe==2.1.3
    # via jinja2
matplotlib==3.7.2
    # via
    #   -r requirements.in
    #   seaborn
mdit-py-plugins==0.3.5
    # via myst-parser
mdurl==0.1.2
    # via markdown-it-py
mmh3==4.0.1
    # via howso-engine-api
myst-parser==1.0.0
    # via -r requirements.in
numpy==1.25.2
    # via
    #   -r requirements.in
    #   contourpy
    #   howso-engine-api
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
packaging==23.1
    # via
    #   deprecation
    #   matplotlib
    #   pydata-sphinx-theme
    #   sphinx
pandas==2.1.0
    # via
    #   howso-engine-api
    #   seaborn
pillow==10.0.0
    # via matplotlib
pycparser==2.21
    # via cffi
pydata-sphinx-theme==0.13.3
    # via -r requirements.in
pygments==2.16.1
    # via
    #   accessible-pygments
    #   pydata-sphinx-theme
    #   rich
    #   sphinx
pyjwt==2.8.0
    # via howso-engine-api
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   faker
    #   howso-openapi-client
    #   matplotlib
    #   pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   howso-engine-api
    #   myst-parser
requests==2.31.0
    # via
    #   sphinx
    #   sphinxcontrib-images
rich==13.5.2
    # via howso-engine-api
scikit-learn==1.3.0
    # via howso-engine-api
scipy==1.11.2
    # via scikit-learn
seaborn==0.12.2
    # via -r requirements.in
semantic-version==2.10.0
    # via howso-engine-api
six==1.16.0
    # via
    #   howso-openapi-client
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.5
    # via beautifulsoup4
sphinx==6.2.1
    # via
    #   -r requirements.in
    #   myst-parser
    #   pydata-sphinx-theme
    #   sphinx-copybutton
    #   sphinx-design
    #   sphinx-gallery
    #   sphinxcontrib-applehelp
    #   sphinxcontrib-devhelp
    #   sphinxcontrib-htmlhelp
    #   sphinxcontrib-images
    #   sphinxcontrib-qthelp
    #   sphinxcontrib-serializinghtml
sphinx-copybutton==0.5.2
    # via -r requirements.in
sphinx-design==0.4.1
    # via -r requirements.in
sphinx-gallery==0.12.2
    # via -r requirements.in
sphinxcontrib-applehelp==1.0.7
    # via sphinx
sphinxcontrib-devhelp==1.0.5
    # via sphinx
sphinxcontrib-htmlhelp==2.0.4
    # via sphinx
sphinxcontrib-images==0.9.4
    # via -r requirements.in
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.6
    # via sphinx
sphinxcontrib-serializinghtml==1.1.9
    # via sphinx
threadpoolctl==3.2.0
    # via scikit-learn
typing-extensions==4.7.1
    # via
    #   howso-engine-api
    #   pydata-sphinx-theme
tzdata==2023.3
    # via pandas
urllib3==1.26.16
    # via
    #   howso-engine-api
    #   howso-openapi-client
    #   requests

# The following packages are considered to be unsafe in a requirements file:
setuptools==68.1.2
    # via howso-openapi-client
